Add files via upload
Fortunately, it’s a lot easier to grasp what a "tensor" is in a practical sense and how we can manipulate them in data science but the real questions is what are Tensors really ? what does the name tensor tells us , what was their motivation for creating it ? This repository contains the history , motivation and manipulations of tensors. The Notebook consists # Creating Tensors # Tensor Shapes #Tensor Datatypes #TensorBroadcasting # Moving To GPu #Changing the number of dimensions and so on.